requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# Data processing with version pins
numpy==1.24.3
pandas==1.5.3

# Google AI and environment
python-dotenv==1.0.0
google-generativeai==0.3.1

# HTTP client
httpx==0.25.0

# Add these new dependencies
jinja2==3.1.2
aiofiles==23.2.1

